requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
dnspython==2.3.0
docopt==0.6.2
email-validator==1.3.1
Flask==2.2.2
Flask-Login==0.6.2
Flask-WTF==1.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.30
mysql-connector-repackaged==0.3.1
numpy==1.24.2
packaging==23.0
pandas==1.5.3
pipreqs==0.4.11
plotly==5.14.0
protobuf==3.20.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.2
six==1.16.0
tenacity==8.2.2
urllib3==1.26.15
Werkzeug==2.2.2
WTForms==3.0.1
yarg==0.1.9